django_project/requirements.txt
django>=5.0.0
pyserial>=3.5
numpy>=1.20.0 